refactor: 인증 서버에 JWT 발급 로직 생성

+++ auth-service/src/main/java/fast/campus/authservice/repository/user/UserJpaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserJpaRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserJpaRepository extends JpaRepository<UserEntity, Long> {
     Optional<UserEntity> findUserEntityByEmail(String userId);
 }
